texocr_app/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow